+++ hd_backend/models/usuario.model.php
@@ -131,15 +131,38 @@ class Usuario
         try {
             $con = new ClaseConectar();
             $con = $con->ProcedimientoParaConectar();
-            $cadena = "DELETE FROM `usuario` WHERE `idUsuario`= $idUsuario";
-            if (mysqli_query($con, $cadena)) {
-                return 1;
-            } else {
-                return $con->error;
-            }
+
+             // Verificar si existen relaciones con el usuario
+             $query = "SELECT COUNT(*) as total FROM ticket WHERE idUsuario = $idUsuario";
+             $result = mysqli_query($con, $query);
+             $row = mysqli_fetch_assoc($result);
+
+             $query1 = "SELECT COUNT(*) as total1 FROM agente WHERE idUsuario = $idUsuario";
+             $result1 = mysqli_query($con, $query1);
+             $row1 = mysqli_fetch_assoc($result1);
+     
+             if ($row['total'] > 0 || $row1['total1'] > 0) {
+                     // devolver un mensaje de error
+                     return [
+                         'status' => 'error',
+                         'message' => 'No se puede eliminar el Usuario, ya se ha vinculado con Tickets o Agentes'
+                     ];
+                 } else {
+                    $cadena = "DELETE FROM `usuario` WHERE `idUsuario`= $idUsuario";
+                    if (mysqli_query($con, $cadena)) {
+                        return 1;
+                    } else {
+                        return [
+                            'status' => 'error',
+                            'message' => 'Error al intentar eliminar el usuario.'
+                        ];
+                    }
+                }
         } catch (Exception $th) {
-            http_response_code(500);
-            return $th->getMessage();
+            return [
+                'status' => 'error',
+                'message' => $th->getMessage()
+            ];
         } finally {
             $con->close();
         }
